+++ docs/CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Version 0.13.0dev (unreleased)
+------------------------------
+
+* Add support for managing remote git repositories (Andreas Kaiser)
+* Add intersphinx to docs (Andreas Kaiser)
+
 Version 0.12.0 (2013-03-04)
 ---------------------------